refactor(app): give the paint frame's two contract types their own module
`DrawFrame` and `AxisChips` move to `pane/draw_frame.rs`, so `draw_chart.rs`
and `layer_painters.rs` both read from it and neither reads from the other:
the second bug pass pointed out that the pair imported each other, a cycle
between children of one module that the cycle guard cannot see. The
baseline's provenance note now names `drawing_gestures.rs` (800, untouched)
as the largest sibling and `draw_chart.rs` as the largest new one.
