+++ crates/app/src/pane/layer_painters.rs
@@ -29,7 +29,7 @@ use crate::theme;
 use crate::toolrail::Tool;
 use quantick_orderflow::engine::VisibleOrderflow;
 
-use super::draw_chart::DrawFrame;
+use super::draw_frame::{AxisChips, DrawFrame};
 use super::tape_switch::TAPE_SWITCH_RESERVED_PX;
 use super::{
     ChartPane, DrawPass, PaneChrome, PointerCompass, PriceAxisLevel, draw_live_chip, live_chip_rect,
@@ -46,17 +46,6 @@ const EMPTY_VIEW_FONT_SIZE: f32 = 16.0;
 /// side of it. Derived from the lane rather than fixed, so widening the lane
 /// widens the candle instead of leaving a wider gap around the same sliver.
 const SIDEBAR_BODY_FRAC: f32 = 0.35;
-
-/// What one frame's axes stand aside for, as `axis_claims` decides it.
-///
-/// Three fields with names rather than a tuple: the two claim lists are the
-/// same type, and a tuple would let the price axis's chips and the time
-/// strip's be swapped by a `let` that still compiles.
-pub(super) struct AxisChips {
-    pub(super) compass: Option<PointerCompass>,
-    pub(super) price: pointer_compass::AxisClaims,
-    pub(super) time: pointer_compass::AxisClaims,
-}
 
 impl ChartPane {
     /// What the axes stand aside for this frame: the pointer compass, the
